Style pod creation flow, update redux state for pod creation
